refactor(project structure): Some changes due to organizing the code's parts have been applied to maintain them better over time.
